requirements.txt
python-telegram-bot==13.7
requests==2.32.3
python-dotenv==1.0.1
urllib3==1.26.15
standard-imghdr